+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # ack-angular - Change Log
 All notable changes to this project will be documented here.
 
-## [0.5.5] - 2017-07-11
+## [0.5.5] - 2017-07-18
 - AckApi has ApiError EventEmitter
 - Updated ack-css-boot
+- UrlVars provider
+- WindowService provider
 
 ## [0.5.3] - 2017-07-10
 ### Enhanced